+++ src/calendar/dot/index.js
@@ -1,10 +1,9 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import {View} from 'react-native';
 import styleConstructor from './style';
-import PropTypes from 'prop-types';
 
 const Dot = ({theme, isMarked, isDisabled, dotColor, isToday, isSelected}) => {
-
   const style = styleConstructor(theme);
   const dotStyle = [style.dot];
 
@@ -28,9 +27,7 @@ const Dot = ({theme, isMarked, isDisabled, dotColor, isToday, isSelected}) => {
     }
   }
 
-  return (
-    <View style={dotStyle}/>
-  );
+  return <View style={dotStyle}/>;
 };
 
 export default Dot;