+++ CHANGELOG.md
@@ -388,3 +388,36 @@
 
 ## [1.406.0] - 2020-11-24
 - Unify library code - this.styles to this.style.
+
+## [1.407.0] - 2020-11-26
+### Fix
+- Calendar - fix for 'disableArrowLeft', 'disableArrowRight' update (PR #1309).
+
+## [1.490.0] - 2020-11-29
+### Added
+- AgendaList - allow passing 'renderSectionHeader' prop (PR #1306).
+
+## [1.491.0] - 2020-11-29
+### Added
+- Agenda - add 'showOnlySelectedDayItems' prop (PR #1317).
+
+## [1.492.0] - 2020-11-29
+### Fix
+- example - fix formatting.
+
+## [1.493.0] - 2020-11-29
+### Fix
+- CalendarList - Forwarding missing 'disabledDaysIndexes' prop from CalendarListItem to Calendar (PR #1323).
+
+## [1.494.0] - 2020-11-29
+- Code cleanup.
+
+## [1.494.0] - 2020-11-29
+### Added
+- why-did-you-render setup.
+
+## [1.495.0] - 2020-12-01
+- Edit README. 
+
+## [1.496.0] - 2020-12-02
+- AgendaList, ExpandableCalendar, WeekCalendar - adding a note to wrap with CalendarProvider.